Move worker scope models to the worker API models
